an important error was fixed
we have shortened the links that are in the "bildiklerim" section
and in this way we solved the responsivity problem for mobile devices.
Because the links were causing the overflow because of their text length
and this overflow situation causes the design problem for mobile devices.
This problem was fixed and now this site is compatible with almost
all devices.