a problematic function removed
a problematical function we create with javascript codes that triggered
when a scroll event occurs was removed. This function formerly was used to
give some effect to the menu buttons depending on the scroll event.
But it had not run properly. Because of that this function and
codes related to this function have been removed.